基于服务方法级别的服务限流，针对服务实例很流，及一个服务多个运行实例很流

+++ apics/src/main/java/cn/jmicro/api/codec/JDataInput.java
@@ -60,8 +60,8 @@ public class JDataInput implements DataInput {
 		return Message.readUnsignedShort(buf);
 	}
 	
-	public int readUnsignedInt() throws IOException{
-		return (int)Message.readUnsignedInt(this.buf);
+	public long readUnsignedInt() throws IOException{
+		return Message.readUnsignedInt(this.buf);
 	}
 
 	@Override
